requirements.txt
dvc
mlflow
torch
torchvision
pandas
scikit-learn
numpy
Pillow